pyproject.toml
[project]
name = "snowflake-learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "prefect>=3.0",
    "snowflake-connector-python>=3.7",
    "minio>=7.2",
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
]

[tool.pytest.ini_options]
pythonpath = ["."]
